+++ packages/tailwind/src/tailwind-presets.ts
@@ -1,7 +1,8 @@
-import * as twDefaultTheme from "tailwindcss/defaultTheme"
 import { ThemeConfig } from "tailwindcss/types/config"
-
 import { defaultTheme } from "@kaizen/design-tokens"
+import { kzHeight } from "./kz-height"
+import { kzSpacing } from "./kz-spacing"
+import { kzWidth } from "./kz-width"
 
 export type KaizenTailwindTheme = Partial<ThemeConfig>
 export interface KaizenTailwindPreset {
@@ -17,32 +18,28 @@ export const kaizenTailwindTheme: KaizenTailwindTheme = {
     ...defaultTheme.color,
     ...defaultTheme.dataViz,
   },
-  spacing: {
-    ...defaultTheme.spacing,
-    ...twDefaultTheme.spacing,
-  },
+  spacing: kzSpacing,
   boxShadow: {
     none: "none",
     sm: defaultTheme.shadow.small.boxShadow,
     lg: defaultTheme.shadow.large.boxShadow,
   },
   borderRadius: {
-    none: "0px",
     default: "7px",
+    none: "0px",
     "focus-ring": "10px",
     full: "100%",
   },
+
   borderWidth: {
-    "width-none": "0px",
-    "width-default": "2px",
-    "width-focus-ring": "2px",
+    "w-none": "0px",
+    "w-default": "2px",
+    "w-focus-ring": "2px",
   },
-  // how do we want to handle our focus rings?
-  // tailwind used a `ring` property
   borderColor: {
-    "color-solid": `${defaultTheme.border.solid.borderColor}`,
-    "color-transparent": `${defaultTheme.border.borderless.borderColor}`,
-    "color-focus-ring": defaultTheme.color.blue[600],
+    "default-color": `${defaultTheme.border.solid.borderColor}`,
+    transparent: `${defaultTheme.border.borderless.borderColor}`,
+    "focus-ring": defaultTheme.color.blue[600],
     ...defaultTheme.color,
   },
   fontFamily: {
@@ -52,82 +49,76 @@ export const kaizenTailwindTheme: KaizenTailwindTheme = {
     "family-display": [`${defaultTheme.typography.display0.fontFamily}`],
   },
   fontSize: {
-    "size-display": `${defaultTheme.typography.display0.fontSize}`,
-    "size-heading-1": `${defaultTheme.typography.heading1.fontSize}`,
-    "size-heading-2": `${defaultTheme.typography.heading2.fontSize}`,
-    "size-heading-3": `${defaultTheme.typography.heading3.fontSize}`,
-    "size-heading-4": `${defaultTheme.typography.heading4.fontSize}`,
-    "size-heading-5": `${defaultTheme.typography.heading5.fontSize}`,
-    "size-heading-6": `${defaultTheme.typography.heading6.fontSize}`,
-    "size-paragraph": `${defaultTheme.typography.paragraphBody.fontSize}`,
-    "size-paragraph-lede": `${defaultTheme.typography.paragraphIntroLede.fontSize}`,
-    "size-paragraph-sm": `${defaultTheme.typography.paragraphSmall.fontSize}`,
-    "size-paragraph-xs": `${defaultTheme.typography.paragraphExtraSmall.fontSize}`,
-    "size-data-lg": `${defaultTheme.typography.dataLarge.fontSize}`,
-    "size-data-md": `${defaultTheme.typography.dataMedium.fontSize}`,
-    "size-data-sm": `${defaultTheme.typography.dataSmall.fontSize}`,
-    "size-data-units-lg": `${defaultTheme.typography.dataLargeUnits.fontSize}`,
-    "size-data-units-md": `${defaultTheme.typography.dataMediumUnits.fontSize}`,
-    "size-data-units-sm": `${defaultTheme.typography.dataSmallUnits.fontSize}`,
+    display: `${defaultTheme.typography.display0.fontSize}`,
+    "heading-1": `${defaultTheme.typography.heading1.fontSize}`,
+    "heading-2": `${defaultTheme.typography.heading2.fontSize}`,
+    "heading-3": `${defaultTheme.typography.heading3.fontSize}`,
+    "heading-4": `${defaultTheme.typography.heading4.fontSize}`,
+    "heading-5": `${defaultTheme.typography.heading5.fontSize}`,
+    "heading-6": `${defaultTheme.typography.heading6.fontSize}`,
+    paragraph: `${defaultTheme.typography.paragraphBody.fontSize}`,
+    "paragraph-lede": `${defaultTheme.typography.paragraphIntroLede.fontSize}`,
+    "paragraph-sm": `${defaultTheme.typography.paragraphSmall.fontSize}`,
+    "paragraph-xs": `${defaultTheme.typography.paragraphExtraSmall.fontSize}`,
+    "data-lg": `${defaultTheme.typography.dataLarge.fontSize}`,
+    "data-md": `${defaultTheme.typography.dataMedium.fontSize}`,
+    "data-sm": `${defaultTheme.typography.dataSmall.fontSize}`,
+    "data-units-lg": `${defaultTheme.typography.dataLargeUnits.fontSize}`,
+    "data-units-md": `${defaultTheme.typography.dataMediumUnits.fontSize}`,
+    "data-units-sm": `${defaultTheme.typography.dataSmallUnits.fontSize}`,
   },
-  // font weights could have specific tokens (current don't)
   fontWeight: {
     "weight-paragraph": "400",
     "weight-paragraph-bold": "600",
     "weight-data": "700",
     "weight-heading": "700",
     "weight-display": "800",
-    "weight-normal": "400",
-    "weight-medium": "500",
-    "weight-semibold": "600",
-    "weight-bold": "700",
-    "weight-extrabold": "800",
   },
   lineHeight: {
-    "height-display": `${defaultTheme.typography.display0.lineHeight}`,
-    "height-heading-1": `${defaultTheme.typography.heading1.lineHeight}`,
-    "height-heading-2": `${defaultTheme.typography.heading2.lineHeight}`,
-    "height-heading-3": `${defaultTheme.typography.heading3.lineHeight}`,
-    "height-heading-4": `${defaultTheme.typography.heading4.lineHeight}`,
-    "height-heading-5": `${defaultTheme.typography.heading5.lineHeight}`,
-    "height-heading-6": `${defaultTheme.typography.heading6.lineHeight}`,
-    "height-paragraph": `${defaultTheme.typography.paragraphBody.lineHeight}`,
-    "height-paragraph-lede": `${defaultTheme.typography.paragraphIntroLede.lineHeight}`,
-    "height-paragraph-sm": `${defaultTheme.typography.paragraphSmall.lineHeight}`,
-    "height-paragraph-xs": `${defaultTheme.typography.paragraphExtraSmall.lineHeight}`,
-    "height-data-lg": `${defaultTheme.typography.dataLarge.lineHeight}`,
-    "height-data-md": `${defaultTheme.typography.dataMedium.lineHeight}`,
-    "height-data-sm": `${defaultTheme.typography.dataSmall.lineHeight}`,
-    "height-data-units-lg": `${defaultTheme.typography.dataLargeUnits.lineHeight}`,
-    "height-data-units-md": `${defaultTheme.typography.dataMediumUnits.lineHeight}`,
-    "height-data-units-sm": `${defaultTheme.typography.dataSmallUnits.lineHeight}`,
+    display: `${defaultTheme.typography.display0.lineHeight}`,
+    "heading-1": `${defaultTheme.typography.heading1.lineHeight}`,
+    "heading-2": `${defaultTheme.typography.heading2.lineHeight}`,
+    "heading-3": `${defaultTheme.typography.heading3.lineHeight}`,
+    "heading-4": `${defaultTheme.typography.heading4.lineHeight}`,
+    "heading-5": `${defaultTheme.typography.heading5.lineHeight}`,
+    "heading-6": `${defaultTheme.typography.heading6.lineHeight}`,
+    paragraph: `${defaultTheme.typography.paragraphBody.lineHeight}`,
+    "paragraph-lede": `${defaultTheme.typography.paragraphIntroLede.lineHeight}`,
+    "paragraph-sm": `${defaultTheme.typography.paragraphSmall.lineHeight}`,
+    "paragraph-xs": `${defaultTheme.typography.paragraphExtraSmall.lineHeight}`,
+    "data-lg": `${defaultTheme.typography.dataLarge.lineHeight}`,
+    "data-md": `${defaultTheme.typography.dataMedium.lineHeight}`,
+    "data-sm": `${defaultTheme.typography.dataSmall.lineHeight}`,
+    "data-units-lg": `${defaultTheme.typography.dataLargeUnits.lineHeight}`,
+    "data-units-md": `${defaultTheme.typography.dataMediumUnits.lineHeight}`,
+    "data-units-sm": `${defaultTheme.typography.dataSmallUnits.lineHeight}`,
   },
   letterSpacing: {
-    "spacing-normal": "normal",
+    "letter-spacing-normal": "normal",
   },
   // A mix of layout styles
-  maxWidth: {
-    ...twDefaultTheme.maxWidth,
-    content: defaultTheme.layout.contentMaxWidth,
-    "content-with-side": defaultTheme.layout.contentMaxWidthWithSidebar,
-  },
-  width: {
-    ...twDefaultTheme.width,
-    "content-side-margin": defaultTheme.layout.contentSideMargin,
-  },
-  height: {
-    ...twDefaultTheme.height,
-    "mobile-actions-drawer-height":
-      defaultTheme.layout.mobileActionsDrawerHeight,
-    "navigation-bar-height": defaultTheme.layout.navigationBarHeight,
-  },
-  // we will probably want to review viewport sizing
+  maxWidth: kzWidth,
+  width: kzWidth,
+  maxHeight: kzHeight,
+  height: kzHeight,
   screens: {
-    "media-min-md": defaultTheme.layout.breakpoints.medium, // => @media (min-width: 768px) { ... }
-    "media-min-lg": defaultTheme.layout.breakpoints.large, // => @media (min-width: 1080px) { ... }
-    "media-max-md": { max: defaultTheme.layout.breakpoints.medium }, // => @media (max-width: 768px) { ... }
-    "media-max-lg": { max: defaultTheme.layout.breakpoints.large }, // => @media (min-width: 1080px) { ... }
+    md: defaultTheme.layout.breakpoints.medium, // => @media (min-width: 768px) { ... }
+    lg: defaultTheme.layout.breakpoints.large, // => @media (min-width: 1080px) { ... }
+    "md-max": { max: defaultTheme.layout.breakpoints.medium }, // => @media (max-width: 768px) { ... }
+    "lg-max": { max: defaultTheme.layout.breakpoints.large }, // => @media (min-width: 1080px) { ... }
   },
+  // TODO: These work in TW Play but not in our Storybook config.
+  //        Add these when we get them working in Storybook.
+  // ringWidth: {
+  //   "default-width": `${defaultTheme.border.focusRing.borderWidth}`,
+  // },
+  // ringOffsetWidth: {
+  //   "default-width": "2px",
+  // },
+  // ringColor: {
+  //   "default-color": `${defaultTheme.color.blue[500]}`,
+  // },
+  // ringOffsetColor: {},
 }
 
 export const Preset: KaizenTailwindPreset = {