+++ packages/tailwind/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.1.3...@kaizen/tailwind@0.2.0) (2022-12-15)
+
+
+### Features
+
+* update tokens presets ([#3165](https://github.com/cultureamp/kaizen-design-system/issues/3165)) ([ac963ef](https://github.com/cultureamp/kaizen-design-system/commit/ac963ef6bf5c30c345fc7f0911089b2ed4b15a49))
+
+
+
+
+
+## [0.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.1.2...@kaizen/tailwind@0.1.3) (2022-12-15)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
 ## [0.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.1.1...@kaizen/tailwind@0.1.2) (2022-12-09)
 
 **Note:** Version bump only for package @kaizen/tailwind